+++ api/fiber/README.md
@@ -1,0 +1,9 @@
+#API FIBER
+
+Esta api é um exemplo utilizando o framework fiber [fiber doc](https://docs.gofiber.io/)
+
+O objetivo é apresentar de forma simples os recursos essenciais para o desenvolvimento de uma API.
+
+### GET
+
+### POST
