+++ api/fiber/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/cors"
 	"github.com/gofiber/fiber/v2/middleware/logger"
-	"github.com/jeffotoni/goexample/api/fiber/crypt"
+	"github.com/jeffotoni/goexample/api/fiber/handler"
 )
 
 func main() {
@@ -28,27 +28,7 @@ func main() {
 		Output:     os.Stdout,
 	}))
 
-	app.Get("/healthz", Healthz)
-	app.Post("/auth/check", Check)
+	app.Get("/healthz", handler.Healthz)
+	app.Post("/auth/check", handler.Check)
 	app.Listen("0.0.0.0:8080")
-}
-
-func Check(c *fiber.Ctx) error {
-	ID, err := crypt.RandomID()
-	if err != nil {
-		return c.Status(400).SendString(`{"msg":"Bad Request RandomID"}`)
-	}
-	c.Set("ID", ID)
-	return c.Status(200).SendString(`{"msg":"ok", "token":"xxxxxxxxxxxxxxxxxx3993"}`)
-}
-
-func Healthz(c *fiber.Ctx) error {
-	ID, err := crypt.RandomID()
-	if err != nil {
-		return c.Status(400).SendString(`{"msg":"Bad Request RandomID"}`)
-	}
-
-	c.Set("ID", ID)
-	c.Set("Content-Type", "application/json")
-	return c.Status(200).SendString("")
 }
